Ajuste para incluir lista de programas nos detalhes de projetos
